Use depth trimming on clones
When processing actions the way that all the nodes are connected
together makes the cloning of nodes get out of control over time. When
inverting the method / resource relationships we don't need a perfect
clone, it turns out 2 levels of depth are sufficient to get us all the
attributes we need.